+++ app/Models/FieldOfficer.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Field_Officer extends Model
+class FieldOfficer extends Model
 {
     use HasFactory;
